requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.26.0
google-auth==1.27.0
google-cloud==0.34.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
packaging==20.9
proto-plus==1.13.0
protobuf==3.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
